Initialize the implementation of Glaucoma prediction
Utilize pytorch built-in models to implement glaucoma prediction, the
datasetp[1] is already splitted nicely as machine learning ready
form. Still we perform data preprocessing to make sure the image form
fits well into our model and optimize memory usage.

The model we use at the first step is DenseNet121 with default weights,
the result shows a promising result, however we can still try to
different alternatives.

Also we utilize atomic mixed precision training here to speed up the
training procedure, we might want to evaluate the effect of it in the
future.

Signed-off-by: I Hsin Cheng <richard120310@gmail.com>